Feat-next Migración a nextjs (#1179) #patch
* feat: install next

* feat(next): add next config and convert home page to support next

* fix(test): fix test in Alert and Meta components

* fix(cart-modal): fix cart when add products and refactor _app.tsx

* fix(overlay-app): remove unnecesary imports in overñay-cart and _app

* feat(next): add export command and add out build into gitignore

* feat(next): rename export folder to dist

* cicd nextjs (#1165)

* cicd nextjs

* chore: modify next step deploy


* chore: rename branch in cd

* chore: rename env NPM_AUTH_TOKEN

* feat(product): add product page to src/pages

* feat(product-page): change react-router to next/router and fix styles

* feat(search-page): change search page to next and fix usBrandFilters

* fix(useBrandFilters): parse as string

* chore: update category and breadcrumb

* chore: restore legacy breadcrumb

* feat(article): add article page and fix useBrandFilters

* fix(usebrandfilters): fix useBrandFilters to not remove actual filters

* feat(landing): add lading page to next

* chore: update collection

* chore: update not found page

* feat(pages): add ResetPassword, ResetPasswordSentMail, UserRegistered and Login

* fix(eslint): delete console.log from useBrandFilters

* fix(eslint): remove unnecesary import

* fix(reset-passwor-mail-sent): fix router to be called in useEffect

* feat: update account

* feat(checkout): start to change page to next

* test: change react router for next/router

* feat(checkout-order-finalized): change to next support

* fix(images): change path images to public assets

* chore: updte images

* chore: updte validation for order detail

* chore: change for getInitialProps

* chore: image element do have explicit width and height

* fix(imports): change import from images to /assets and remove react-router and react-router-dom

* fix(dynatrace): add preconnect to _document.tsx

* chore: change ci and cd workflows for nextjs support

* refactor: update image loading

* chore: dont change package-lock.json

* chore: dont change package-lock.json

* chore: change ubication dynatrace

* chore: push to main

* chore: push to deploy vercel

* chore: push to deploy vercel

* chore: load images according to the screen size

* chore: add cdn for image svgs in home

* chore: load images with react lazy images

* fake commit

* fake commit

* fix(Fonts): fix fonts on local

* fix(OverlayNav): increase by 1 zIndex

+++ src/@sdk/repository/Repository.ts
@@ -1,8 +1,10 @@
+import { LOCAL_STORAGE_EXISTS } from '../consts';
 import { NamedObservable } from '../helpers';
 import { LocalStorageItems } from './types';
 
 export class Repository extends NamedObservable<LocalStorageItems> {
   protected saveItem(name: LocalStorageItems, item: string | null): void {
+    if (!LOCAL_STORAGE_EXISTS) return;
     if (item) {
       localStorage?.setItem(name, item);
     } else {
@@ -11,12 +13,14 @@ export class Repository extends NamedObservable<LocalStorageItems> {
     this.notifyChange(name, item);
   }
   protected retrieveItem(name: LocalStorageItems): string | null {
+    if (!LOCAL_STORAGE_EXISTS) return null;
     return localStorage?.getItem(name);
   }
   protected saveObject<T extends object>(
     name: LocalStorageItems,
     object: T | null
   ): void {
+    if (!LOCAL_STORAGE_EXISTS) return;
     if (object) {
       localStorage?.setItem(name, JSON.stringify(object));
     } else {
@@ -27,6 +31,7 @@ export class Repository extends NamedObservable<LocalStorageItems> {
   protected retrieveObject<T extends object>(
     name: LocalStorageItems
   ): T | null {
+    if (!LOCAL_STORAGE_EXISTS) return null;
     const item = localStorage?.getItem(name);
     if (item) {
       return JSON.parse(item);
@@ -34,6 +39,7 @@ export class Repository extends NamedObservable<LocalStorageItems> {
     return null;
   }
   protected clearRepositoryStorage() {
+    if (!LOCAL_STORAGE_EXISTS) return;
     localStorage.clear();
   }
 }
